polish(pu): use legal_actions_cython rather than legal_actions_cython_str
